fix: refactor helper class for data tests
