+++ tests/CNAB.Infra.Data.Test/CNAB.Infra.Data.Test.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -22,6 +22,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\CNAB.Domain\CNAB.Domain.csproj" />
     <ProjectReference Include="..\..\src\CNAB.Infra.Data\CNAB.Infra.Data.csproj" />
+    <ProjectReference Include="..\CNAB.TestHelpers\CNAB.TestHelpers.csproj" />
   </ItemGroup>
 
   <ItemGroup>
